+++ fdb-record-layer-core/src/test/java/com/apple/foundationdb/record/provider/foundationdb/query/FDBInQueryTest.java
@@ -946,7 +946,7 @@ class FDBInQueryTest extends FDBRecordStoreQueryTestBase {
                 },
                 rec -> Tuple.from(rec.getNumValueUnique()),
                 reverse,
-                (useCascadesPlanner || omitPrimaryKeyInOrdering) ? TestHelpers::assertDiscardedNone : (cx -> TestHelpers.assertDiscardedAtMost(79, cx))));
+                (isUseCascadesPlanner() || omitPrimaryKeyInOrdering) ? TestHelpers::assertDiscardedNone : (cx -> TestHelpers.assertDiscardedAtMost(79, cx))));
     }
 
     @DualPlannerTest(planner = DualPlannerTest.Planner.CASCADES)
@@ -1324,7 +1324,7 @@ class FDBInQueryTest extends FDBRecordStoreQueryTestBase {
                 },
                 rec -> Tuple.from(rec.getNumValueUnique()),
                 reverse,
-                (useCascadesPlanner || (omitPrimaryKeyInOrdering && inAsOrUnionMode != InAsOrUnionMode.NONE)) ? TestHelpers::assertDiscardedNone : (cx -> TestHelpers.assertDiscardedAtMost(79, cx))));
+                (isUseCascadesPlanner() || (omitPrimaryKeyInOrdering && inAsOrUnionMode != InAsOrUnionMode.NONE)) ? TestHelpers::assertDiscardedNone : (cx -> TestHelpers.assertDiscardedAtMost(79, cx))));
     }
 
     /**