+++ fdb-record-layer-core/src/test/java/com/apple/foundationdb/record/provider/foundationdb/query/FDBPermutedMinMaxQueryTest.java
@@ -175,7 +175,7 @@ class FDBPermutedMinMaxQueryTest extends FDBRecordStoreQueryTestBase {
     @ParameterizedTest
     @BooleanSource
     void selectMaxOrderByFirstGroup(boolean reverse) throws Exception {
-        Assumptions.assumeTrue(useCascadesPlanner);
+        Assumptions.assumeTrue(isUseCascadesPlanner());
         final RecordMetaDataHook hook = metaData -> metaData.addIndex(metaData.getRecordType("MySimpleRecord"), maxUniqueBy2And3());
         complexQuerySetup(hook);
 
@@ -275,7 +275,7 @@ class FDBPermutedMinMaxQueryTest extends FDBRecordStoreQueryTestBase {
     @ParameterizedTest
     @BooleanSource
     void selectMaxByGroupWithOrder(boolean reverse) throws Exception {
-        Assumptions.assumeTrue(useCascadesPlanner);
+        Assumptions.assumeTrue(isUseCascadesPlanner());
         final RecordMetaDataHook hook = metaData -> metaData.addIndex(metaData.getRecordType("MySimpleRecord"), maxUniqueBy2And3());
         complexQuerySetup(hook);
 
@@ -493,7 +493,7 @@ class FDBPermutedMinMaxQueryTest extends FDBRecordStoreQueryTestBase {
     @ParameterizedTest
     @BooleanSource
     void selectMaxGroupByWithPredicateAndOrderByOnMax(boolean reverse) throws Exception {
-        Assumptions.assumeTrue(useCascadesPlanner);
+        Assumptions.assumeTrue(isUseCascadesPlanner());
         final RecordMetaDataHook hook = metaData -> metaData.addIndex(metaData.getRecordType("MySimpleRecord"), maxUniqueBy2And3());
         complexQuerySetup(hook);