+++ fdb-record-layer-lucene/src/test/java/com/apple/foundationdb/record/lucene/LuceneIndexMaintenanceTest.java
@@ -20,13 +20,21 @@
 
 package com.apple.foundationdb.record.lucene;
 
+import com.apple.foundationdb.record.IndexEntry;
 import com.apple.foundationdb.record.LoggableTimeoutException;
 import com.apple.foundationdb.record.RecordCoreException;
+import com.apple.foundationdb.record.RecordCursor;
 import com.apple.foundationdb.record.RecordMetaData;
 import com.apple.foundationdb.record.RecordMetaDataBuilder;
+import com.apple.foundationdb.record.ScanProperties;
 import com.apple.foundationdb.record.TestRecordsGroupedParentChildProto;
+import com.apple.foundationdb.record.TestRecordsTextProto;
 import com.apple.foundationdb.record.logging.KeyValueLogMessage;
 import com.apple.foundationdb.record.logging.LogMessageKeys;
+import com.apple.foundationdb.record.lucene.directory.AgilityContext;
+import com.apple.foundationdb.record.lucene.directory.FDBDirectory;
+import com.apple.foundationdb.record.lucene.directory.FDBDirectoryLockFactory;
+import com.apple.foundationdb.record.lucene.directory.FDBDirectoryWrapper;
 import com.apple.foundationdb.record.metadata.Index;
 import com.apple.foundationdb.record.metadata.JoinedRecordTypeBuilder;
 import com.apple.foundationdb.record.metadata.Key;
@@ -34,18 +42,24 @@ import com.apple.foundationdb.record.metadata.expressions.KeyExpression;
 import com.apple.foundationdb.record.metadata.expressions.ThenKeyExpression;
 import com.apple.foundationdb.record.provider.common.StoreTimer;
 import com.apple.foundationdb.record.provider.foundationdb.FDBRecordContext;
-import com.apple.foundationdb.record.provider.foundationdb.FDBRecordStoreTestBase;
+import com.apple.foundationdb.record.provider.foundationdb.FDBRecordStore;
+import com.apple.foundationdb.record.provider.foundationdb.FDBRecordStoreConcurrentTestBase;
+import com.apple.foundationdb.record.provider.foundationdb.IndexMaintainerState;
+import com.apple.foundationdb.record.provider.foundationdb.IndexMaintenanceFilter;
 import com.apple.foundationdb.record.provider.foundationdb.OnlineIndexer;
 import com.apple.foundationdb.record.provider.foundationdb.properties.RecordLayerPropertyStorage;
 import com.apple.foundationdb.record.query.expressions.Query;
+import com.apple.foundationdb.record.query.plan.QueryPlanner;
+import com.apple.foundationdb.record.util.pair.Pair;
+import com.apple.foundationdb.subspace.Subspace;
 import com.apple.foundationdb.tuple.Tuple;
 import com.apple.test.RandomizedTestUtils;
 import com.apple.test.SuperSlow;
 import com.apple.test.Tags;
 import com.apple.test.TestConfigurationUtils;
-import org.apache.commons.lang3.tuple.Pair;
-import org.junit.jupiter.api.Assertions;
+import org.apache.lucene.store.Lock;
 import org.junit.jupiter.api.Tag;
+import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.Arguments;
 import org.junit.jupiter.params.provider.MethodSource;
@@ -53,6 +67,7 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 import javax.annotation.Nonnull;
+import javax.annotation.Nullable;
 import java.io.IOException;
 import java.time.Instant;
 import java.util.ArrayDeque;
@@ -61,20 +76,36 @@ import java.util.HashMap;
 import java.util.IdentityHashMap;
 import java.util.List;
 import java.util.Map;
+import java.util.Objects;
 import java.util.Random;
+import java.util.concurrent.CountDownLatch;
+import java.util.concurrent.ThreadLocalRandom;
 import java.util.concurrent.TimeUnit;
 import java.util.concurrent.atomic.AtomicInteger;
-import java.util.function.Consumer;
 import java.util.function.Function;
+import java.util.stream.Collectors;
 import java.util.stream.Stream;
+
+import static com.apple.foundationdb.record.lucene.LuceneIndexOptions.INDEX_PARTITION_BY_FIELD_NAME;
+import static com.apple.foundationdb.record.lucene.LuceneIndexOptions.INDEX_PARTITION_HIGH_WATERMARK;
+import static com.apple.foundationdb.record.metadata.Key.Expressions.concat;
+import static com.apple.foundationdb.record.metadata.Key.Expressions.field;
+import static com.apple.foundationdb.record.metadata.Key.Expressions.function;
+import static org.junit.jupiter.api.Assertions.assertEquals;
+import static org.junit.jupiter.api.Assertions.assertFalse;
+import static org.junit.jupiter.api.Assertions.assertNotNull;
+import static org.junit.jupiter.api.Assertions.assertThrows;
 
 /**
  * Tests of the consistency of the Lucene Index.
  */
 @Tag(Tags.RequiresFDB)
-public class LuceneIndexMaintenanceTest extends FDBRecordStoreTestBase {
-
+public class LuceneIndexMaintenanceTest extends FDBRecordStoreConcurrentTestBase {
     private static final Logger LOGGER = LoggerFactory.getLogger(LuceneIndexMaintenanceTest.class);
+
+    public LuceneIndexMaintenanceTest() {
+        super(null);
+    }
 
     static Stream<Arguments> configurationArguments() {
         // This has found situations that should have explicit tests:
@@ -128,7 +159,7 @@ public class LuceneIndexMaintenanceTest extends FDBRecordStoreTestBase {
         final KeyExpression rootExpression = createRootExpression(isGrouped, isSynthetic);
         Index index = addIndex(isSynthetic, rootExpression, options, metaDataBuilder);
         final RecordMetaData metadata = metaDataBuilder.build();
-        Consumer<FDBRecordContext> schemaSetup = context -> createOrOpenRecordStore(context, metadata);
+        Function<FDBRecordContext, Pair<FDBRecordStore, QueryPlanner>> schemaSetup = context -> createOrOpenRecordStore(context, metadata);
 
         final RecordLayerPropertyStorage contextProps = RecordLayerPropertyStorage.newBuilder()
                 .addProp(LuceneRecordContextProperties.LUCENE_REPARTITION_DOCUMENT_COUNT, repartitionCount)
@@ -142,10 +173,8 @@ public class LuceneIndexMaintenanceTest extends FDBRecordStoreTestBase {
 
         explicitMergeIndex(index, contextProps, schemaSetup);
 
-        new LuceneIndexTestValidator(() -> openContext(contextProps), context -> {
-            schemaSetup.accept(context);
-            return recordStore;
-        }).validate(index, ids, repartitionCount, isSynthetic ? "child_str_value:forth" : "text_value:about");
+        new LuceneIndexTestValidator(() -> openContext(contextProps), context -> Objects.requireNonNull(schemaSetup.apply(context).getLeft()))
+                .validate(index, ids, repartitionCount, isSynthetic ? "child_str_value:forth" : "text_value:about");
 
         if (isGrouped) {
             validateDeleteWhere(isSynthetic, repartitionCount, ids, contextProps, schemaSetup, index);
@@ -234,7 +263,7 @@ public class LuceneIndexMaintenanceTest extends FDBRecordStoreTestBase {
         final KeyExpression rootExpression = createRootExpression(isGrouped, isSynthetic);
         Index index = addIndex(isSynthetic, rootExpression, options, metaDataBuilder);
         final RecordMetaData metadata = metaDataBuilder.build();
-        Consumer<FDBRecordContext> schemaSetup = context -> createOrOpenRecordStore(context, metadata);
+        Function<FDBRecordContext, Pair<FDBRecordStore, QueryPlanner>> schemaSetup = context -> createOrOpenRecordStore(context, metadata);
 
         final RecordLayerPropertyStorage contextProps = RecordLayerPropertyStorage.newBuilder()
                 .addProp(LuceneRecordContextProperties.LUCENE_REPARTITION_DOCUMENT_COUNT, repartitionCount)
@@ -255,10 +284,7 @@ public class LuceneIndexMaintenanceTest extends FDBRecordStoreTestBase {
             explicitMergeIndex(index, contextProps, schemaSetup);
         }
 
-        final LuceneIndexTestValidator luceneIndexTestValidator = new LuceneIndexTestValidator(() -> openContext(contextProps), context -> {
-            schemaSetup.accept(context);
-            return recordStore;
-        });
+        final LuceneIndexTestValidator luceneIndexTestValidator = new LuceneIndexTestValidator(() -> openContext(contextProps), context -> Objects.requireNonNull(schemaSetup.apply(context).getLeft()));
         luceneIndexTestValidator.validate(index, ids, repartitionCount, isSynthetic ? "child_str_value:forth" : "text_value:about");
 
         if (isGrouped) {
@@ -325,7 +351,7 @@ public class LuceneIndexMaintenanceTest extends FDBRecordStoreTestBase {
         final KeyExpression rootExpression = createRootExpression(isGrouped, isSynthetic);
         Index index = addIndex(isSynthetic, rootExpression, options, metaDataBuilder);
         final RecordMetaData metadata = metaDataBuilder.build();
-        Consumer<FDBRecordContext> schemaSetup = context -> createOrOpenRecordStore(context, metadata);
+        Function<FDBRecordContext, Pair<FDBRecordStore, QueryPlanner>> schemaSetup = context -> createOrOpenRecordStore(context, metadata);
 
         final RecordLayerPropertyStorage contextProps = RecordLayerPropertyStorage.newBuilder()
                 .addProp(LuceneRecordContextProperties.LUCENE_MERGE_SEGMENTS_PER_TIER, 2.0)
@@ -339,10 +365,10 @@ public class LuceneIndexMaintenanceTest extends FDBRecordStoreTestBase {
         final Map<Tuple, Map<Tuple, Tuple>> ids = new HashMap<>();
         generateDocuments(isGrouped, isSynthetic, minDocumentCount, random, contextProps, schemaSetup, transactionCount, ids);
 
-        final Function<StoreTimer.Wait, Pair<Long, TimeUnit>> oldAsyncToSyncTimeout = fdb.getAsyncToSyncTimeout();
+        final Function<StoreTimer.Wait, org.apache.commons.lang3.tuple.Pair<Long, TimeUnit>> oldAsyncToSyncTimeout = fdb.getAsyncToSyncTimeout();
         AtomicInteger waitCounts = new AtomicInteger();
         try {
-            final Function<StoreTimer.Wait, Pair<Long, TimeUnit>> asyncToSyncTimeout = (wait) -> {
+            final Function<StoreTimer.Wait, org.apache.commons.lang3.tuple.Pair<Long, TimeUnit>> asyncToSyncTimeout = (wait) -> {
                 if (wait.getClass().equals(LuceneEvents.Waits.class) &&
                         // don't have the timeout on FILE_LOCK_CLEAR because that will leave the file lock around,
                         // and the next iteration will fail on that.
@@ -352,9 +378,9 @@ public class LuceneIndexMaintenanceTest extends FDBRecordStoreTestBase {
                         wait != LuceneEvents.Waits.WAIT_LUCENE_FILE_LOCK_SET &&
                         waitCounts.getAndDecrement() == 0) {
 
-                    return Pair.of(1L, TimeUnit.NANOSECONDS);
+                    return org.apache.commons.lang3.tuple.Pair.of(1L, TimeUnit.NANOSECONDS);
                 } else {
-                    return oldAsyncToSyncTimeout == null ? Pair.of(1L, TimeUnit.DAYS) : oldAsyncToSyncTimeout.apply(wait);
+                    return oldAsyncToSyncTimeout == null ? org.apache.commons.lang3.tuple.Pair.of(1L, TimeUnit.DAYS) : oldAsyncToSyncTimeout.apply(wait);
                 }
             };
             for (int i = 0; i < 100; i++) {
@@ -367,7 +393,7 @@ public class LuceneIndexMaintenanceTest extends FDBRecordStoreTestBase {
                     explicitMergeIndex(index, contextProps, schemaSetup);
                     LOGGER.info(KeyValueLogMessage.of("Merge completed",
                             "iteration", i));
-                    Assertions.assertFalse(requireFailure && i < 15, i + " merge should have failed");
+                    assertFalse(requireFailure && i < 15, i + " merge should have failed");
                     success = true;
                 } catch (RecordCoreException e) {
                     final LoggableTimeoutException timeoutException = findTimeoutException(e);
@@ -379,17 +405,15 @@ public class LuceneIndexMaintenanceTest extends FDBRecordStoreTestBase {
                     if (timeoutException == null) {
                         throw e;
                     }
-                    Assertions.assertEquals(1L, timeoutException.getLogInfo().get(LogMessageKeys.TIME_LIMIT.toString()), i + " " + e.getMessage());
-                    Assertions.assertEquals(TimeUnit.NANOSECONDS, timeoutException.getLogInfo().get(LogMessageKeys.TIME_UNIT.toString()), i + " " + e.getMessage());
+                    assertEquals(1L, timeoutException.getLogInfo().get(LogMessageKeys.TIME_LIMIT.toString()), i + " " + e.getMessage());
+                    assertEquals(TimeUnit.NANOSECONDS, timeoutException.getLogInfo().get(LogMessageKeys.TIME_UNIT.toString()), i + " " + e.getMessage());
                 }
                 fdb.setAsyncToSyncTimeout(oldAsyncToSyncTimeout);
                 dbExtension.checkForOpenContexts(); // validate after every loop that we didn't leave any contexts open
                 LOGGER.debug(KeyValueLogMessage.of("Validating",
                         "iteration", i));
-                new LuceneIndexTestValidator(() -> openContext(contextProps), context -> {
-                    schemaSetup.accept(context);
-                    return recordStore;
-                }).validate(index, ids, Integer.MAX_VALUE, isSynthetic ? "child_str_value:forth" : "text_value:about", !success);
+                new LuceneIndexTestValidator(() -> openContext(contextProps), context -> Objects.requireNonNull(schemaSetup.apply(context).getLeft()))
+                        .validate(index, ids, Integer.MAX_VALUE, isSynthetic ? "child_str_value:forth" : "text_value:about", !success);
                 LOGGER.debug(KeyValueLogMessage.of("Done Validating",
                         "iteration", i));
                 dbExtension.checkForOpenContexts(); // just in case the validation code leaks a context
@@ -399,6 +423,326 @@ public class LuceneIndexMaintenanceTest extends FDBRecordStoreTestBase {
             if (LOGGER.isDebugEnabled()) {
                 ids.entrySet().stream().sorted(Map.Entry.comparingByKey())
                         .forEach(entry -> LOGGER.debug(entry.getKey() + ": " + entry.getValue().keySet()));
+            }
+        }
+    }
+
+    // Lock a directory, and commit, then try to update a record, or save a record, or do a search, and assert that nothing
+    // is corrupted (or that the user request fails)
+    @Test
+    void lockCommitThenValidateTest() throws IOException {
+        final Map<String, String> options = Map.of(
+                INDEX_PARTITION_BY_FIELD_NAME, "timestamp",
+                INDEX_PARTITION_HIGH_WATERMARK, String.valueOf(8));
+        Index index = complexPartitionedIndex(options);
+
+        Function<FDBRecordContext, Pair<FDBRecordStore, QueryPlanner>> schemaSetup = context ->
+                LuceneIndexTestUtils.rebuildIndexMetaData(context, path, TestRecordsTextProto.ComplexDocument.getDescriptor().getName(), index, useCascadesPlanner);
+
+        final RecordLayerPropertyStorage contextProps = RecordLayerPropertyStorage.newBuilder()
+                .addProp(LuceneRecordContextProperties.LUCENE_REPARTITION_DOCUMENT_COUNT, 8)
+                .build();
+
+        long timestamp = System.currentTimeMillis();
+        Map<Tuple, Map<Tuple, Tuple>> insertedDocs = new HashMap<>();
+        // save a record
+        createComplexRecords(1, insertedDocs, contextProps, schemaSetup);
+
+        // explicitly lock directory then commit
+        try (FDBRecordContext context = openContext(contextProps)) {
+            FDBRecordStore recordStore = Objects.requireNonNull(schemaSetup.apply(context).getLeft());
+
+            // low-level lock the directory:
+            // subspace for (group 1, partition data subspace, partition 0, file lock subspace)
+            Subspace subspace = recordStore.indexSubspace(index).subspace(Tuple.from(1, LucenePartitioner.PARTITION_DATA_SUBSPACE, 0, FDBDirectory.FILE_LOCK_SUBSPACE));
+            byte[] fileLockKey = subspace.pack(Tuple.from("write.lock"));
+            FDBDirectoryLockFactory lockFactory = new FDBDirectoryLockFactory(null, 10_000);
+
+            Lock testLock = lockFactory.obtainLock(new AgilityContext.NonAgile(context), fileLockKey, "write.lock");
+            testLock.ensureValid();
+            commit(context);
+        }
+
+        // search should work
+        try (FDBRecordContext context = openContext(contextProps)) {
+            FDBRecordStore recordStore = Objects.requireNonNull(schemaSetup.apply(context).getLeft());
+
+            try (RecordCursor<IndexEntry> cursor = recordStore.scanIndex(
+                    index,
+                    LuceneIndexTestValidator.groupedSortedTextSearch(recordStore, index, "text:word", null, 1), null, ScanProperties.FORWARD_SCAN)) {
+                List<Tuple> primaryKeys = cursor.asList()
+                        .join()
+                        .stream()
+                        .map(IndexEntry::getPrimaryKey)
+                        .collect(Collectors.toList());
+                assertEquals(1, primaryKeys.size());
+                assertEquals(Tuple.from(1, 1000L), primaryKeys.get(0));
+            }
+        }
+
+        // create another record, this should fail
+        try (FDBRecordContext context = openContext(contextProps)) {
+            FDBRecordStore recordStore = Objects.requireNonNull(schemaSetup.apply(context).getLeft());
+
+            TestRecordsTextProto.ComplexDocument cd = TestRecordsTextProto.ComplexDocument.newBuilder()
+                    .setGroup(1)
+                    .setDocId(2000L)
+                    .setIsSeen(true)
+                    .setText("A word about what I want to say")
+                    .setTimestamp(timestamp + 2000)
+                    .setHeader(TestRecordsTextProto.ComplexDocument.Header.newBuilder().setHeaderId(1999L))
+                    .build();
+            assertThrows(RecordCoreException.class, () -> recordStore.saveRecord(cd), "Lock failed: already locked by another entity");
+        }
+
+        // validate index
+        new LuceneIndexTestValidator(() -> openContext(contextProps), context -> Objects.requireNonNull(schemaSetup.apply(context).getLeft()))
+                .validate(index, insertedDocs, Integer.MAX_VALUE, "text:about", false);
+    }
+
+
+    // A chaos test of two threads, one constantly trying to merge, and one trying to update a record, or save a new record or do a search.
+    // At the end the index should be validated for consistency.
+    @Test
+    void chaosMergeAndUpdateTest() throws InterruptedException, IOException {
+        final Map<String, String> options = Map.of(
+                INDEX_PARTITION_BY_FIELD_NAME, "timestamp",
+                INDEX_PARTITION_HIGH_WATERMARK, String.valueOf(100));
+        Index index = complexPartitionedIndex(options);
+
+        Function<FDBRecordContext, Pair<FDBRecordStore, QueryPlanner>> schemaSetup = context ->
+                LuceneIndexTestUtils.rebuildIndexMetaData(context, path, TestRecordsTextProto.ComplexDocument.getDescriptor().getName(), index, useCascadesPlanner);
+
+        assertNotNull(index);
+
+        final RecordLayerPropertyStorage contextProps = RecordLayerPropertyStorage.newBuilder()
+                .addProp(LuceneRecordContextProperties.LUCENE_REPARTITION_DOCUMENT_COUNT, 8)
+                .build();
+
+        final int countReps = 20;
+
+        long timestamp = System.currentTimeMillis();
+        Map<Tuple, Map<Tuple, Tuple>> insertedDocs = new HashMap<>();
+        CountDownLatch firstRecordInserted = new CountDownLatch(1);
+        Thread inserter = new Thread(() -> {
+            for (int i = 0; i < countReps; i++) {
+                // create a record then query
+                try (FDBRecordContext context = openContext(contextProps)) {
+                    FDBRecordStore recordStore = Objects.requireNonNull(schemaSetup.apply(context).getLeft());
+                    recordStore.getIndexDeferredMaintenanceControl().setAutoMergeDuringCommit(false);
+                    TestRecordsTextProto.ComplexDocument cd = TestRecordsTextProto.ComplexDocument.newBuilder()
+                            .setGroup(1)
+                            .setDocId(i + 1000L)
+                            .setIsSeen(true)
+                            .setText("A word about what I want to say")
+                            .setTimestamp(timestamp + i)
+                            .setHeader(TestRecordsTextProto.ComplexDocument.Header.newBuilder().setHeaderId(1000L - i))
+                            .build();
+                    try {
+                        final Tuple primaryKey = recordStore.saveRecord(cd).getPrimaryKey();
+
+                        try (RecordCursor<IndexEntry> cursor = recordStore.scanIndex(
+                                index,
+                                LuceneIndexTestValidator.groupedSortedTextSearch(recordStore, index, "text:word", null, 1), null, ScanProperties.FORWARD_SCAN)) {
+                            List<IndexEntry> matches = cursor.asList().join();
+                            assertFalse(matches.isEmpty());
+                        }
+
+                        commit(context);
+                        insertedDocs.computeIfAbsent(Tuple.from(1), k -> new HashMap<>()).put(primaryKey, Tuple.from(timestamp + i));
+                        // after first record is committed, signal to merger record to start attempting to merge
+                        firstRecordInserted.countDown();
+                    } catch (Exception e) {
+                        // commit failed due to conflict with other thread. Continue trying to create docs.
+                        LOGGER.debug("couldn't commit for key {}", (1000L + i));
+                    }
+                }
+            }
+        });
+
+        Thread merger = new Thread(() -> {
+            // wait till first record is committed before attempting to merge
+            try {
+                firstRecordInserted.await();
+            } catch (InterruptedException interruptedException) {
+                Thread.currentThread().interrupt();
+                return;
+            }
+
+            // busy merge
+            for (int i = 0; i < countReps; i++) {
+                explicitMergeIndex(index, contextProps, schemaSetup);
+            }
+        });
+
+        inserter.start();
+        merger.start();
+
+        inserter.join();
+        merger.join();
+
+        // validate index is sane
+        new LuceneIndexTestValidator(() -> openContext(contextProps), context -> Objects.requireNonNull(schemaSetup.apply(context).getLeft()))
+                .validate(index, insertedDocs, Integer.MAX_VALUE, "text:about", false);
+    }
+
+    // A test where there are multiple threads trying to do merges. At the end the index should be validated for consistency.
+    @Test
+    void multipleConcurrentMergesTest() throws IOException, InterruptedException {
+        final Map<String, String> options = Map.of(
+                INDEX_PARTITION_BY_FIELD_NAME, "timestamp",
+                INDEX_PARTITION_HIGH_WATERMARK, String.valueOf(100));
+
+        Index index = complexPartitionedIndex(options);
+
+        Function<FDBRecordContext, Pair<FDBRecordStore, QueryPlanner>> schemaSetup = context ->
+                LuceneIndexTestUtils.rebuildIndexMetaData(context, path, TestRecordsTextProto.ComplexDocument.getDescriptor().getName(), index, useCascadesPlanner);
+        final RecordLayerPropertyStorage contextProps = RecordLayerPropertyStorage.newBuilder()
+                .addProp(LuceneRecordContextProperties.LUCENE_REPARTITION_DOCUMENT_COUNT, 8)
+                .build();
+
+        final int countReps = 20;
+        final int threadCount = 10;
+
+        Map<Tuple, Map<Tuple, Tuple>> insertedDocs = new HashMap<>();
+
+        // create a bunch of docs
+        createComplexRecords(countReps, insertedDocs, contextProps, schemaSetup);
+
+        final CountDownLatch readyToMerge = new CountDownLatch(1);
+        final CountDownLatch doneMerging = new CountDownLatch(threadCount);
+
+        for (int i = 0; i < threadCount; i++) {
+            new Thread(() -> {
+                try {
+                    readyToMerge.await();
+                    explicitMergeIndex(index, contextProps, schemaSetup);
+                } catch (InterruptedException interruptedException) {
+                    Thread.currentThread().interrupt();
+                } finally {
+                    doneMerging.countDown();
+                }
+            }).start();
+        }
+        readyToMerge.countDown();
+        doneMerging.await();
+
+        new LuceneIndexTestValidator(() -> openContext(contextProps), context -> Objects.requireNonNull(schemaSetup.apply(context).getLeft()))
+                .validate(index, insertedDocs, Integer.MAX_VALUE, "text:about", false);
+    }
+
+    static Stream<Arguments> mergeLosesLockTest() {
+        return Stream.concat(
+                Stream.of( 65).map(Arguments::of), // fixed 65% lock failure rate
+                RandomizedTestUtils.randomArguments(random -> Arguments.of(random.nextInt(101)))); //  0-100%
+    }
+
+    // A test of what lucene does when a merge loses its lock
+    @ParameterizedTest
+    @MethodSource
+    void mergeLosesLockTest(int failurePercentage) throws IOException {
+        final Map<String, String> options = Map.of(
+                INDEX_PARTITION_BY_FIELD_NAME, "timestamp",
+                INDEX_PARTITION_HIGH_WATERMARK, String.valueOf(200));
+        Index index = complexPartitionedIndex(options);
+
+        Function<FDBRecordContext, Pair<FDBRecordStore, QueryPlanner>> schemaSetup = context ->
+                LuceneIndexTestUtils.rebuildIndexMetaData(context, path, TestRecordsTextProto.ComplexDocument.getDescriptor().getName(), index, useCascadesPlanner);
+
+        final RecordLayerPropertyStorage contextProps = RecordLayerPropertyStorage.newBuilder()
+                .addProp(LuceneRecordContextProperties.LUCENE_REPARTITION_DOCUMENT_COUNT, 8)
+                .addProp(LuceneRecordContextProperties.LUCENE_MERGE_SEGMENTS_PER_TIER, 2.0)
+                .addProp(LuceneRecordContextProperties.LUCENE_AGILE_COMMIT_TIME_QUOTA, 1) // 1ms
+                .build();
+
+        final int docCount = 100;
+        Map<Tuple, Map<Tuple, Tuple>> insertedDocs = new HashMap<>();
+        // create a bunch of docs
+        createComplexRecords(docCount, insertedDocs, contextProps, schemaSetup);
+
+        // try a couple of times
+        for (int l = 0; l < 2; l++) {
+            try (FDBRecordContext context = openContext(contextProps)) {
+                FDBRecordStore recordStore = Objects.requireNonNull(schemaSetup.apply(context).getLeft());
+
+                // directory key for group 1/partition 0
+                Tuple directoryKey = Tuple.from(1, LucenePartitioner.PARTITION_DATA_SUBSPACE, 0);
+                IndexMaintainerState state = new IndexMaintainerState(recordStore, index, IndexMaintenanceFilter.NORMAL);
+
+                // custom test directory that returns a lucene lock that's never valid (Lock.ensureValid() throws IOException)
+                FDBDirectory fdbDirectory = new InvalidLockTestFDBDirectory(recordStore.indexSubspace(index).subspace(directoryKey), context, options, failurePercentage);
+                FDBDirectoryWrapper fdbDirectoryWrapper = new FDBDirectoryWrapper(state, fdbDirectory, directoryKey, 1, AgilityContext.agile(context, 1L, 1L));
+
+                final var fieldInfos = LuceneIndexExpressions.getDocumentFieldDerivations(state.index, state.store.getRecordMetaData());
+                LuceneAnalyzerCombinationProvider indexAnalyzerSelector = LuceneAnalyzerRegistryImpl.instance().getLuceneAnalyzerCombinationProvider(state.index, LuceneAnalyzerType.FULL_TEXT, fieldInfos);
+
+                assertThrows(IOException.class, () -> fdbDirectoryWrapper.mergeIndex(indexAnalyzerSelector.provideIndexAnalyzer(""), new Exception()), "invalid lock");
+                commit(context);
+            }
+        }
+
+        // validate that the index is still sane
+        new LuceneIndexTestValidator(() -> openContext(contextProps), context -> Objects.requireNonNull(schemaSetup.apply(context).getLeft()))
+                .validate(index, insertedDocs, Integer.MAX_VALUE, "text:about", false);
+    }
+
+    /**
+     * a test FDBDirectory class that returns a {@link Lock} that is not valid.
+     */
+    static class InvalidLockTestFDBDirectory extends FDBDirectory {
+        private final int percentFailure;
+
+        public InvalidLockTestFDBDirectory(@Nonnull Subspace subspace,
+                                           @Nonnull FDBRecordContext context,
+                                           @Nullable Map<String, String> indexOptions,
+                                           final int percentFailure) {
+            super(subspace, context, indexOptions);
+            this.percentFailure = percentFailure;
+        }
+
+        @Override
+        @Nonnull
+        public Lock obtainLock(@Nonnull final String lockName) throws IOException {
+            final Lock lock = super.obtainLock(lockName);
+            return new Lock() {
+                @Override
+                public void close() throws IOException {
+                    lock.close();
+                }
+
+                @Override
+                public void ensureValid() throws IOException {
+                    // 0 <= 0-99 < 100
+                    if (ThreadLocalRandom.current().nextInt(100) < percentFailure) {
+                        throw new IOException("invalid lock");
+                    } else {
+                        lock.ensureValid();
+                    }
+                }
+            };
+        }
+    }
+
+    private void createComplexRecords(int count,
+                                      Map<Tuple, Map<Tuple, Tuple>> insertedKeys,
+                                      RecordLayerPropertyStorage contextProps,
+                                      Function<FDBRecordContext, Pair<FDBRecordStore, QueryPlanner>> schemaSetup) {
+        long timestamp = System.currentTimeMillis();
+        for (int i = 0; i < count; i++) {
+            try (FDBRecordContext context = openContext(contextProps)) {
+                FDBRecordStore recordStore = Objects.requireNonNull(schemaSetup.apply(context).getLeft());
+                recordStore.getIndexDeferredMaintenanceControl().setAutoMergeDuringCommit(false);
+                TestRecordsTextProto.ComplexDocument cd = TestRecordsTextProto.ComplexDocument.newBuilder()
+                        .setGroup(1)
+                        .setDocId(i + 1000L)
+                        .setIsSeen(true)
+                        .setText("A word about what I want to say")
+                        .setTimestamp(timestamp + i)
+                        .setHeader(TestRecordsTextProto.ComplexDocument.Header.newBuilder().setHeaderId(1000L - i))
+                        .build();
+                Tuple primaryKey = recordStore.saveRecord(cd).getPrimaryKey();
+                insertedKeys.computeIfAbsent(Tuple.from(1), k -> new HashMap<>()).put(primaryKey, Tuple.from(timestamp));
+                commit(context);
             }
         }
     }
@@ -425,13 +769,14 @@ public class LuceneIndexMaintenanceTest extends FDBRecordStoreTestBase {
         return null;
     }
 
-    @Nonnull
-    private void generateDocuments(final boolean isGrouped, final boolean isSynthetic,
-                               final int minDocumentCount, final Random random,
-                               final RecordLayerPropertyStorage contextProps,
-                               final Consumer<FDBRecordContext> schemaSetup,
-                               final int transactionCount,
-                               final Map<Tuple, Map<Tuple, Tuple>> ids) {
+    private void generateDocuments(final boolean isGrouped,
+                                   final boolean isSynthetic,
+                                   final int minDocumentCount,
+                                   final Random random,
+                                   final RecordLayerPropertyStorage contextProps,
+                                   final Function<FDBRecordContext, Pair<FDBRecordStore, QueryPlanner>> schemaSetup,
+                                   final int transactionCount,
+                                   final Map<Tuple, Map<Tuple, Tuple>> ids) {
         final long start = Instant.now().toEpochMilli();
         int i = 0;
         while (i < transactionCount ||
@@ -443,14 +788,14 @@ public class LuceneIndexMaintenanceTest extends FDBRecordStoreTestBase {
                         .orElse(0) < minDocumentCount) {
             final int docCount = random.nextInt(10) + 1;
             try (FDBRecordContext context = openContext(contextProps)) {
-                schemaSetup.accept(context);
+                FDBRecordStore recordStore = Objects.requireNonNull(schemaSetup.apply(context).getLeft());
                 recordStore.getIndexDeferredMaintenanceControl().setAutoMergeDuringCommit(false);
                 for (int j = 0; j < docCount; j++) {
                     final int group = isGrouped ? random.nextInt(random.nextInt(10) + 1) : 0; // irrelevant if !isGrouped
                     final Tuple groupTuple = isGrouped ? Tuple.from(group) : Tuple.from();
                     final int countInGroup = ids.computeIfAbsent(groupTuple, key -> new HashMap<>()).size();
                     long timestamp = start + countInGroup + random.nextInt(20) - 5;
-                    final Tuple primaryKey = saveRecords(isSynthetic, group, countInGroup, timestamp, random);
+                    final Tuple primaryKey = saveRecords(recordStore, isSynthetic, group, countInGroup, timestamp, random);
                     ids.computeIfAbsent(groupTuple, key -> new HashMap<>()).put(primaryKey, Tuple.from(timestamp).addAll(primaryKey));
                 }
                 commit(context);
@@ -470,24 +815,41 @@ public class LuceneIndexMaintenanceTest extends FDBRecordStoreTestBase {
         return metaDataBuilder;
     }
 
-    private void validateDeleteWhere(final boolean isSynthetic, final int repartitionCount, final Map<Tuple, Map<Tuple, Tuple>> ids, final RecordLayerPropertyStorage contextProps, final Consumer<FDBRecordContext> schemaSetup, final Index index) throws IOException {
+    @Nonnull
+    public static Index complexPartitionedIndex(final Map<String, String> options) {
+        return new Index("Complex$partitioned",
+                concat(function(LuceneFunctionNames.LUCENE_TEXT, field("text")),
+                        function(LuceneFunctionNames.LUCENE_SORTED, field("timestamp"))).groupBy(field("group")),
+                LuceneIndexTypes.LUCENE,
+                options);
+    }
+
+    private void validateDeleteWhere(final boolean isSynthetic,
+                                     final int repartitionCount,
+                                     final Map<Tuple, Map<Tuple, Tuple>> ids,
+                                     final RecordLayerPropertyStorage contextProps,
+                                     final Function<FDBRecordContext, Pair<FDBRecordStore, QueryPlanner>> schemaSetup,
+                                     final Index index) throws IOException {
         final List<Tuple> groups = List.copyOf(ids.keySet());
         for (final Tuple group : groups) {
             try (FDBRecordContext context = openContext(contextProps)) {
-                schemaSetup.accept(context);
+                FDBRecordStore recordStore = Objects.requireNonNull(schemaSetup.apply(context).getLeft());
                 recordStore.deleteRecordsWhere(Query.field("group").equalsValue(group.getLong(0)));
                 context.commit();
             }
             ids.remove(group);
-            new LuceneIndexTestValidator(() -> openContext(contextProps), context -> {
-                schemaSetup.accept(context);
-                return recordStore;
-            }).validate(index, ids, repartitionCount, isSynthetic ? "child_str_value:forth" : "text_value:about");
+            new LuceneIndexTestValidator(() -> openContext(contextProps), context -> Objects.requireNonNull(schemaSetup.apply(context).getLeft()))
+                    .validate(index, ids, repartitionCount, isSynthetic ? "child_str_value:forth" : "text_value:about");
         }
     }
 
     @Nonnull
-    private Tuple saveRecords(final boolean isSynthetic, final int group, final int countInGroup, final long timestamp, final Random random) {
+    private Tuple saveRecords(final FDBRecordStore recordStore,
+                              final boolean isSynthetic,
+                              final int group,
+                              final int countInGroup,
+                              final long timestamp,
+                              final Random random) {
         var parent = TestRecordsGroupedParentChildProto.MyParentRecord.newBuilder()
                 .setGroup(group)
                 .setRecNo(1001L + countInGroup)
@@ -574,9 +936,11 @@ public class LuceneIndexMaintenanceTest extends FDBRecordStoreTestBase {
     }
 
 
-    private void explicitMergeIndex(Index index, RecordLayerPropertyStorage contextProps, Consumer<FDBRecordContext> schemaSetup) {
+    private void explicitMergeIndex(Index index,
+                                    RecordLayerPropertyStorage contextProps,
+                                    Function<FDBRecordContext, Pair<FDBRecordStore, QueryPlanner>> schemaSetup) {
         try (FDBRecordContext context = openContext(contextProps)) {
-            schemaSetup.accept(context);
+            FDBRecordStore recordStore = Objects.requireNonNull(schemaSetup.apply(context).getLeft());
             try (OnlineIndexer indexBuilder = OnlineIndexer.newBuilder()
                     .setRecordStore(recordStore)
                     .setIndex(index)
@@ -585,5 +949,9 @@ public class LuceneIndexMaintenanceTest extends FDBRecordStoreTestBase {
                 indexBuilder.mergeIndex();
             }
         }
+    }
+
+    protected RecordLayerPropertyStorage.Builder addDefaultProps(final RecordLayerPropertyStorage.Builder props) {
+        return super.addDefaultProps(props).addProp(LuceneRecordContextProperties.LUCENE_INDEX_COMPRESSION_ENABLED, true);
     }
 }